added data export

+++ app/tasks.py
@@ -2,6 +2,7 @@ from datetime import date
 import os
 import sys
 import json
+import boto3
 from rq import get_current_job
 from flask import render_template, url_for
 from flask_login import current_user
@@ -35,42 +36,55 @@ def _set_task_progress(progress):
         db.session.commit()
 
 
-def delete_account(uid, ext):
+def account_export(uid, ext, delete=False):
     try:
         _set_task_progress(0)
         user = User.query.get(uid)
-        
-        az_path_base = f'{user.id}/exports/'
-        filename = get_zip(user, ext)
-        az_path = f'{az_path_base}lurnby-export.zip'
-        s3.upload_file(
-            Bucket = bucket,
+        if ext != 'none':
+            print('exporting')
+            if os.environ.get('DEV'):
+                az_path_base = f'staging/{user.id}/exports/'
+            else:
+                az_path_base = f'{user.id}/exports/'
+            filename = get_zip(user, ext)
+            az_path = f'{az_path_base}lurnby-export.zip'
+            s3.upload_file(
+                Bucket = bucket,
 
-            Filename=filename,
-            Key=az_path
-            )
-        location = s3.get_bucket_location(Bucket=bucket)['LocationConstraint']
-        #url = "https://s3-%s.amazonaws.com/%s/%s" % (location, bucket, az_path)
-        url = s3.generate_presigned_url('get_object', Params = {'Bucket': bucket, 'Key': az_path}, ExpiresIn = 604800)
+                Filename=filename,
+                Key=az_path
+                )
+            location = s3.get_bucket_location(Bucket=bucket)['LocationConstraint']
+            #url = "https://s3-%s.amazonaws.com/%s/%s" % (location, bucket, az_path)
+            url = s3.generate_presigned_url('get_object', Params = {'Bucket': bucket, 'Key': az_path}, ExpiresIn = 604800)
 
 
-        send_email('[Lurnby] Your exported data',
-                sender=app.config['ADMINS'][0], recipients=[user.email],
-                text_body=render_template('email/export_highlights.txt', url=url, user=user),
-                html_body=render_template('email/export_highlights.html', url=url, user=user),
-                sync=True)
+            send_email('[Lurnby] Your exported data',
+                    sender=app.config['ADMINS'][0], recipients=[user.email],
+                    text_body=render_template('email/export_highlights.txt', url=url, user=user),
+                    html_body=render_template('email/export_highlights.html', url=url, user=user),
+                    sync=True)
+        if delete:
+            if os.environ.get('DEV'):
+                az_path_base = f'staging/{user.id}/'
+            else:
+                az_path_base = f'{user.id}/'
 
-        for tag in user.tags.all():
-            db.session.delete(tag)
-        for highlight in user.highlights.all():
-            db.session.delete(highlight)
-        for topic in user.topics.all():
-            db.session.delete(topic)
-        for article in user.articles.all():
-            db.session.delete(article)
-        _set_task_progress(100)
-        db.session.delete(user)
-        db.session.commit()
+            az = boto3.resource('s3')
+            buck = az.Bucket(bucket)
+            buck.objects.filter(Prefix=az_path_base).delete()
+
+            for tag in user.tags.all():
+                db.session.delete(tag)
+            for highlight in user.highlights.all():
+                db.session.delete(highlight)
+            for topic in user.topics.all():
+                db.session.delete(topic)
+            for article in user.articles.all():
+                db.session.delete(article)
+            _set_task_progress(100)
+            db.session.delete(user)
+            db.session.commit()
         
     except:
         app.logger.error('Unhandled exception', exc_info=sys.exc_info())
@@ -92,7 +106,11 @@ def export_highlights(user, highlights, source, ext):
         if not os.path.isdir(path):
             os.mkdir(path)
 
-        az_path_base = f'{user.id}/exports/'
+        if os.environ.get('DEV'):
+            az_path_base = f'staging/{user.id}/exports/'
+        else:
+            az_path_base = f'{user.id}/exports/'
+
 
         if ext == 'txt':
             if source == 'article':
